Добавлены проверки доступа к редактированию и удалению книг в приложении books_app

+++ books_app/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 
 from .models import Book
 from .forms import BookForm
@@ -30,6 +32,7 @@ def view_detail_book(request, book_id):
     return render(request, "books/detail_book.html", context=context)
 
 
+@login_required(login_url='users:login')
 def add_book_view(request):
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES)
@@ -43,8 +46,15 @@ def add_book_view(request):
     return render(request, 'books/add_book.html', {'form': form})
 
 
+@login_required(login_url='users:login')
 def edit_book_view(request, book_id):
     book = get_object_or_404(Book, id=book_id)
+
+    if book.seller != request.user:
+        raise PermissionDenied(
+            "У вас нет прав на редактирование этой книги."
+    )
+
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES, instance=book)
         if form.is_valid():
@@ -55,9 +65,15 @@ def edit_book_view(request, book_id):
     return render(request, 'books/add_book.html', {'form': form})
 
 
+@login_required(login_url='users:login')
 def delete_book_view(request, book_id):
     book = get_object_or_404(Book, id=book_id)
-    print(book)
+
+    if book.seller != request.user:
+        raise PermissionDenied(
+            "У вас нет прав на редактирование этой книги."
+        )
+
     if request.method == 'POST':
         book.delete()
         return redirect('books:index')
